+++ ssl/ssl_lib.c
@@ -1260,7 +1260,8 @@ char *SSL_get_shared_ciphers(const SSL *s, char *buf, int len)
     for (i = 0; i < sk_SSL_CIPHER_num(sk); i++) {
         c = sk_SSL_CIPHER_value(sk, i);
         end = buf + curlen;
-        if (strlcat(buf, c->name, len) >= len || (curlen = strlcat(buf, ":", len)) >= len) {
+        if (strlcat(buf, c->name, len) >= (size_t)len ||
+           (curlen = strlcat(buf, ":", len)) >= (size_t)len) {
             /* remove truncated cipher from list */
             *end = '\0';
             break;
